Added window abstraction for events

+++ src/vec.c
@@ -1,7 +1,7 @@
 #include "int_types.h"
 #include "vec.h"
 
-Vec vec_new(s16 x, s16 y) {
+Vec vec_new(s32 x, s32 y) {
 	Vec value = {
 		x,
 		y
